Make Greeting & Exit modular
Greeting & Exiting was done within the main()
function, which may not be good practice especially
by exposing how the system is ran.

Created private functions greet() and exit() to
abstract away the details on how to greet and exit
the function.

This allows users to merely use the functions,
providing a clean interface and allow for easier
editing in the future.