Improve code modularity
Code repetition for the initial "ChatterBox: " as
well as the lack of modularity for processing user input.

Created a processUserInput() function in order to abstract away
how user input is processed, and created a name variable to store
the initial output of "ChatterBox: ".

This helps to improve code modularity and improve readability,
maintainability and reusability.