Add support for mark & unmark
Lack of support for marking tasks in list.

Added support for marking tasks as complete / incomplete.
List of items done:
* Moved functionality for processing commands into a class
called CommandProcessor to improve modularity and extensibility.
* All commands are now processed via the CommandProcessor class,
with the exception of adding tasks since there is no command for now.
* Made the generic Storage class extend Task in order to access Task methods.
* Runnable interface is created in order to utilize lambda functions
for the commands.
* Scanner is passed into the CommandProcessor class methods in order
to prevent undefined behaviour from multiple scanners for I/O.