Add support for various Tasks
Previous CommandProcessor relied on Task class objects.

With the addition of new types of Tasks, we needed to
modify the CommandProcessor and ChatterBox class to support
it. CommandProcessor now takes in Storage<Task> as opposed
to Storage<T> for Runnable objects since we are only storing
Task objects in the storage object.