Account for missing input for some commands
When user inputs commands alone, some commands would result in errors

Added NoSuchElementException catches such that error checking is done.

For user input, a newline character is inserted in order to mimic how
user input was done on the CLI. This reduces the amount of code that
needs to be changed.